Create hooks directory
The old `conf.d` directory was there to allow the repository to be used
as a fisherman plugin, but this was always a bit of a lie: the user
would still need to manually distribute the `random-colors` binary.

The new plan is to make the binary the one thing you need. It can hook
itself into your shell.