Store entire filepath in `Location` type
We were turning the Location type into a full filepath in an
increasingly large number of places. Each of those calls is effectfull,
in that we need to ask the outside world where the configuration
directory is so we can then calculate the full file path.

If we store the full filepath in the `Location` to begin with then
there's no need to convert it later.