Shorter documentation strings.
Strings beyond the wrap length nimpretty causes it to behave in a
non-idempotent way: Consecutive runs of nimpretty on the same file will
move the string a line down for each run.

This is a bug in nempretty. Until it is fixed the best thing to do is
not to use such long strings. Or not use nimpretty, but I'm not prepared
to give it up yet.