Store location hashes
These location hashes will be used as the filenames in which we store
color schema information, replacing a base64 encoding used before. The
base64 encoding has as the disadvantage that it gets longer as the
location paths get longer. This could result in some cases in the color
scheme location file names becoming longer than what was supported by
the operating system, resulting in an error.